libvmaf/ciede: avoid cross-platform mismatches
moves a few calculations to double precision
error should be below the logging precision now

+++ libvmaf/src/feature/ciede.c
@@ -125,7 +125,7 @@ static float get_h_prime(const float x, const float y)
 {
     if ((x == 0.0) && (y == 0.0))
         return 0.0;
-    float hue_angle = atan2f(x, y);
+    float hue_angle = atan2(x, y);
     if (hue_angle < 0.0)
         hue_angle += 2. * M_PI;
     return hue_angle;
@@ -156,10 +156,10 @@ static float get_upcase_h_bar_prime(const float h_prime_1,
 static float get_upcase_t(const float upcase_h_bar_prime)
 {
     return 1.0 -
-           0.17 * cosf(upcase_h_bar_prime - M_PI / 6.0) +
-           0.24 * cosf(2.0 * upcase_h_bar_prime) +
-           0.32 * cosf(3.0 * upcase_h_bar_prime + M_PI / 30.0) -
-           0.20 * cosf(4.0 * upcase_h_bar_prime - 7.0 * M_PI / 20.0);
+           0.17 * cos(upcase_h_bar_prime - M_PI / 6.0) +
+           0.24 * cos(2.0 * upcase_h_bar_prime) +
+           0.32 * cos(3.0 * upcase_h_bar_prime + M_PI / 30.0) -
+           0.20 * cos(4.0 * upcase_h_bar_prime - 7.0 * M_PI / 20.0);
 }
 
 static float radians_to_degrees(const float radians)
@@ -219,7 +219,7 @@ static float ciede2000(LABColor color_1, LABColor color_2, KSubArgs ksub)
     const float h_prime_2 = get_h_prime(color_2.b, a_prime_2);
     const float delta_h_prime = get_delta_h_prime(c1, c2, h_prime_1, h_prime_2);
     const float delta_upcase_h_prime =
-            2.0 * sqrtf(c_prime_1 * c_prime_2) * sinf(delta_h_prime / 2.0);
+            2.0 * sqrt(c_prime_1 * c_prime_2) * sin(delta_h_prime / 2.0);
     const float upcase_h_bar_prime =
         get_upcase_h_bar_prime(h_prime_1, h_prime_2);
     const float upcase_t = get_upcase_t(upcase_h_bar_prime);
@@ -233,32 +233,32 @@ static float ciede2000(LABColor color_1, LABColor color_2, KSubArgs ksub)
                 pow(hue, 2) + r_sub_t * chroma * hue);
 }
 
-static float pow_2_4(float x)
+static double pow_2_4(double x)
 {
-    return powf(x, 2.4);
+    return pow(x, 2.4);
 }
 
-static float rgb_to_xyz_map(float c)
+static double rgb_to_xyz_map(double c)
 {
     if (c > 10. / 255.) {
-        const float A = 0.055;
-        const float D = 1.0 / 1.055;
+        const double A = 0.055;
+        const double D = 1.0 / 1.055;
         return pow_2_4((c + A) * D);
     } else {
-        const float D = 1.0 / 12.92;
+        const double D = 1.0 / 12.92;
         return (c * D);
     }
 }
 
-static float cbrt_approx(float c)
+static double cbrt_approx(double c)
 {
-    return powf(c, 1.0 / 3.0);
+    return pow(c, 1.0 / 3.0);
 }
 
-static float xyz_to_lab_map(float c)
+static float xyz_to_lab_map(double c)
 {
-    const float KAPPA = 24389.0 / 27.0;
-    const float EPSILON = 216.0 / 24389.0;
+    const double KAPPA = 24389.0 / 27.0;
+    const double EPSILON = 216.0 / 24389.0;
 
     if (c > EPSILON) {
         return cbrt_approx(c);
@@ -267,28 +267,28 @@ static float xyz_to_lab_map(float c)
     }
 }
 
-static LABColor get_lab_color(float y, float u, float v, unsigned bpc)
+static LABColor get_lab_color(double y, double u, double v, unsigned bpc)
 {
-    const float scale = 1 << (bpc - 8);
+    const double scale = 1 << (bpc - 8);
 
     y = (y - 16.  * scale) * (1. / (219. * scale));
     u = (u - 128. * scale) * (1. / (224. * scale));
     v = (v - 128. * scale) * (1. / (224. * scale));
 
     // Assumes BT.709
-    float r = y + 1.28033 * v;
-    float g = y - 0.21482 * u - 0.38059 * v;
-    float b = y + 2.12798 * u;
+    double r = y + 1.28033 * v;
+    double g = y - 0.21482 * u - 0.38059 * v;
+    double b = y + 2.12798 * u;
 
     r = rgb_to_xyz_map(r);
     g = rgb_to_xyz_map(g);
     b = rgb_to_xyz_map(b);
 
-    float x = r * 0.4124564390896921 + g * 0.357576077643909 +
+    double x = r * 0.4124564390896921 + g * 0.357576077643909 +
               b * 0.18043748326639894;
           y = r * 0.21267285140562248 + g * 0.715152155287818 +
               b * 0.07217499330655958;
-    float z = r * 0.019333895582329317 + g * 0.119192025881303 +
+    double z = r * 0.019333895582329317 + g * 0.119192025881303 +
               b * 0.9503040785363677;
 
     x = xyz_to_lab_map(x * (1.0 / 0.95047));
